# services/api/docs/swagger/swagger.yaml
components:
...
  schemas:
...
      GetResult_any.any.any_: { properties: {}, type: object },
...
      Coin:
        {
          properties:
            {
              btcPrice: { type: string },
              24hVolume: { type: string },
              coinrankingUrl: { type: string },
              lowVolume: { type: boolean },
              sparkline: { items: { type: string }, type: array },
              rank: { type: number, format: double },
              change: { type: string },
              tier: { type: number, format: double },
              listedAt: { type: number, format: double },
              price: { type: string },
              marketCap: { type: string },
              iconUrl: { type: string },
              color: { type: string },
              name: { type: string },
              symbol: { type: string },
              uuid: { type: string },
            },
          required:
            [
              btcPrice,
              24hVolume,
              coinrankingUrl,
              lowVolume,
              sparkline,
              rank,
              change,
              tier,
              listedAt,
              price,
              marketCap,
              iconUrl,
              color,
              name,
              symbol,
              uuid,
            ],
          type: object,
        },
      TimePeriod:
        { type: string, enum: [1h, 3h, 12h, 24h, 7d, 30d, 3m, 1y, 3y, 5y] },
      Tier: { type: string, enum: ['1', '2', '3'] },
      OrderBy:
        { type: string, enum: [price, marketCap, 24hVolume, change, listedAt] },
      runtime.Types.DefaultSelection_CurrencyPayload_:
        {
          allOf:
            [
              { $ref: '#/components/schemas/GetResult_any.any.any_' },
              { properties: {}, type: object },
            ],
        },
      Currency:
        {
          $ref: '#/components/schemas/runtime.Types.DefaultSelection_CurrencyPayload_',
          description: 'Model Currency',
        },
      ForexSource: { type: string, enum: [fixer, frankfurter] },
      ForexHistory:
        {
          properties:
            {
              to: { type: number, format: double },
              from: { type: number, format: double },
              date: { type: string },
            },
          required: [to, from, date],
          type: object,
        },
...
      NewsArticle:
        {
          properties:
            {
              content: { type: string },
              publishedAt: { type: string },
              urlToImage: { type: string },
              url: { type: string },
              description: { type: string },
              title: { type: string },
              author: { type: string },
              source:
                {
                  properties: { name: { type: string }, id: { type: string } },
                  required: [name, id],
                  type: object,
                },
            },
          required:
            [
              content,
              publishedAt,
              urlToImage,
              url,
              description,
              title,
              author,
              source,
            ],
          type: object,
        },
      NewsCountry:
        {
          type: string,
          enum:
            [
              ae,
              ar,
              at,
              au,
              be,
              bg,
              br,
              ca,
              ch,
              cn,
              co,
              cu,
              cz,
              de,
              eg,
              fr,
              gb,
              gr,
              hk,
              hu,
              id,
              ie,
              il,
              in,
              it,
              jp,
              kr,
              lt,
              lv,
              ma,
              mx,
              my,
              ng,
              nl,
              no,
              nz,
              ph,
              pl,
              pt,
              ro,
              rs,
              ru,
              sa,
              se,
              sg,
              si,
              sk,
              th,
              tr,
              tw,
              ua,
              us,
              ve,
              za,
            ],
        },
      NewsCategory:
        {
          type: string,
          enum:
            [
              business,
              entertainment,
              general,
              health,
              science,
              sports,
              technology,
            ],
        },
      NewsLanguage:
        {
          type: string,
          enum: [ar, de, en, es, fr, he, it, nl, no, pt, ru, sv, ud, zh],
        },
      NewsSource:
        {
          properties:
            {
              country: { $ref: '#/components/schemas/NewsCountry' },
              language: { $ref: '#/components/schemas/NewsLanguage' },
              category: { $ref: '#/components/schemas/NewsCategory' },
              url: { type: string },
              description: { type: string },
              name: { type: string },
              id: { type: string },
            },
          required: [country, language, category, url, description, name, id],
          type: object,
        },
...
      YouTubeCategory:
        {
          properties:
            {
              channelId: { type: string },
              title: { type: string },
              id: { type: string },
            },
          required: [channelId, title, id],
          type: object,
        },
      Thumbnail:
        {
          properties:
            {
              height: { type: number, format: double },
              width: { type: number, format: double },
              url: { type: string },
            },
          required: [height, width, url],
          type: object,
        },
      Thumbnails:
        {
          properties:
            {
              maxres: { $ref: '#/components/schemas/Thumbnail' },
              standard: { $ref: '#/components/schemas/Thumbnail' },
              high: { $ref: '#/components/schemas/Thumbnail' },
              medium: { $ref: '#/components/schemas/Thumbnail' },
              default: { $ref: '#/components/schemas/Thumbnail' },
            },
          required: [maxres, standard, high, medium, default],
          type: object,
        },
      YouTubeVideo:
        {
          properties:
            {
              thumbnails: { $ref: '#/components/schemas/Thumbnails' },
              channelTitle: { type: string },
              channelId: { type: string },
              description: { type: string },
              title: { type: string },
              id: { type: string },
            },
          required:
            [thumbnails, channelTitle, channelId, description, title, id],
          type: object,
        },
...
  securitySchemes: {}
info:
  title: '@world/api'
  version: 0.0.1
  description: 'World API'
  license: { name: 'GNU GPL V3.0' }
  contact: { name: 'hieudoanm ', email: hieumdoan@gmail.com }
openapi: 3.0.0
paths:
...
  /crypto/coins:
    {
      get:
        {
          operationId: GetCoins,
          responses:
            {
              '200':
                {
                  description: '',
                  content:
                    {
                      application/json:
                        {
                          schema:
                            {
                              items: { $ref: '#/components/schemas/Coin' },
                              type: array,
                            },
                        },
                    },
                },
            },
          tags: [Crypto],
          security: [],
          parameters:
            [
              {
                in: query,
                name: timePeriod,
                required: false,
                schema: { $ref: '#/components/schemas/TimePeriod' },
              },
              {
                in: query,
                name: tier,
                required: false,
                schema: { $ref: '#/components/schemas/Tier' },
              },
              {
                in: query,
                name: orderBy,
                required: false,
                schema: { $ref: '#/components/schemas/OrderBy' },
              },
              {
                in: query,
                name: limit,
                required: false,
                schema: { default: 500, format: double, type: number },
              },
            ],
        },
    }
  '/crypto/coins/{id}':
    {
      get:
        {
          operationId: GetCoin,
          responses:
            {
              '200':
                {
                  description: '',
                  content:
                    {
                      application/json:
                        { schema: { $ref: '#/components/schemas/Coin' } },
                    },
                },
            },
          tags: [Crypto],
          security: [],
          parameters:
            [{ in: path, name: id, required: true, schema: { type: string } }],
        },
    }
...
  /currencies/rates:
    {
      get:
        {
          operationId: GetRates,
          responses:
            {
              '200':
                {
                  description: '',
                  content:
                    {
                      application/json:
                        {
                          schema:
                            {
                              items:
                                {
                                  properties:
                                    {
                                      rate: { type: number, format: double },
                                      code: { type: string },
                                    },
                                  required: [rate, code],
                                  type: object,
                                },
                              type: array,
                            },
                        },
                    },
                },
            },
          tags: [Currencies],
          security: [],
          parameters:
            [
              {
                in: query,
                name: amount,
                required: false,
                schema: { default: 1, format: double, type: number },
              },
              {
                in: query,
                name: base,
                required: false,
                schema: { default: EUR, type: string },
              },
              {
                in: query,
                name: to,
                required: false,
                schema: { default: '', type: string },
              },
              {
                in: query,
                name: source,
                required: false,
                schema: { $ref: '#/components/schemas/ForexSource' },
              },
            ],
        },
    }
  /currencies/history:
    {
      get:
        {
          operationId: GetHistory,
          responses:
            {
              '200':
                {
                  description: Ok,
                  content:
                    {
                      application/json:
                        {
                          schema:
                            {
                              items:
                                { $ref: '#/components/schemas/ForexHistory' },
                              type: array,
                            },
                        },
                    },
                },
            },
          tags: [Currencies],
          security: [],
          parameters:
            [
              {
                in: query,
                name: amount,
                required: false,
                schema: { default: 1, format: double, type: number },
              },
              {
                in: query,
                name: days,
                required: false,
                schema: { default: 7, format: double, type: number },
              },
              {
                in: query,
                name: from,
                required: false,
                schema: { default: EUR, type: string },
              },
              {
                in: query,
                name: to,
                required: false,
                schema: { default: USD, type: string },
              },
            ],
        },
    }
  '/currencies/{code}':
    {
      get:
        {
          operationId: GetCurrency,
          responses:
            {
              '200':
                {
                  description: Ok,
                  content:
                    {
                      application/json:
                        { schema: { $ref: '#/components/schemas/Currency' } },
                    },
                },
            },
          tags: [Currencies],
          security: [],
          parameters:
            [
              {
                in: path,
                name: code,
                required: true,
                schema: { type: string },
              },
            ],
        },
    }
...
  /google/trends:
    {
      get:
        {
          operationId: GetTrends,
          responses:
            {
              '200':
                {
                  description: '',
                  content:
                    {
                      application/json:
                        {
                          schema:
                            {
                              items:
                                {
                                  allOf:
                                    [
                                      {
                                        $ref: '#/components/schemas/GetResult_any.any.any_',
                                      },
                                      { properties: {}, type: object },
                                    ],
                                },
                              type: array,
                            },
                        },
                    },
                },
            },
          tags: [Google],
          security: [],
          parameters:
            [
              {
                in: query,
                name: countryCode,
                required: false,
                schema: { default: '', type: string },
              },
            ],
        },
    }
  /google/ranks:
    {
      get:
        {
          operationId: GetRanks,
          responses:
            {
              '200':
                {
                  description: '',
                  content:
                    {
                      application/json:
                        {
                          schema:
                            {
                              items:
                                {
                                  properties:
                                    {
                                      count: { type: number, format: double },
                                      query: { type: string },
                                      rank: { type: number, format: double },
                                    },
                                  required: [count, query, rank],
                                  type: object,
                                },
                              type: array,
                            },
                        },
                    },
                },
            },
          tags: [Google],
          security: [],
          parameters:
            [
              {
                in: query,
                name: offset,
                required: false,
                schema: { default: 0, format: double, type: number },
              },
              {
                in: query,
                name: limit,
                required: false,
                schema: { default: 10, format: double, type: number },
              },
            ],
        },
    }
...
  /news/headlines:
    {
      get:
        {
          operationId: GetHeadlines,
          responses:
            {
              '200':
                {
                  description: '',
                  content:
                    {
                      application/json:
                        {
                          schema:
                            {
                              items:
                                { $ref: '#/components/schemas/NewsArticle' },
                              type: array,
                            },
                        },
                    },
                },
            },
          tags: [News],
          security: [],
          parameters:
            [
              {
                in: query,
                name: country,
                required: false,
                schema: { $ref: '#/components/schemas/NewsCountry' },
              },
              {
                in: query,
                name: category,
                required: false,
                schema: { $ref: '#/components/schemas/NewsCategory' },
              },
              {
                in: query,
                name: sources,
                required: false,
                schema: { default: '', type: string },
              },
              {
                in: query,
                name: q,
                required: false,
                schema: { default: '', type: string },
              },
              {
                in: query,
                name: pageSize,
                required: false,
                schema: { default: 10, format: double, type: number },
              },
            ],
        },
    }
  /news/sources:
    {
      get:
        {
          operationId: GetSources,
          responses:
            {
              '200':
                {
                  description: '',
                  content:
                    {
                      application/json:
                        {
                          schema:
                            {
                              items:
                                { $ref: '#/components/schemas/NewsSource' },
                              type: array,
                            },
                        },
                    },
                },
            },
          tags: [News],
          security: [],
          parameters:
            [
              {
                in: query,
                name: category,
                required: false,
                schema: { $ref: '#/components/schemas/NewsCategory' },
              },
              {
                in: query,
                name: language,
                required: false,
                schema: { $ref: '#/components/schemas/NewsLanguage' },
              },
              {
                in: query,
                name: country,
                required: false,
                schema: { $ref: '#/components/schemas/NewsCountry' },
              },
            ],
        },
    }
...
  /youtube/categories:
    {
      get:
        {
          operationId: GetVideoCategories,
          responses:
            {
              '200':
                {
                  description: '',
                  content:
                    {
                      application/json:
                        {
                          schema:
                            {
                              items:
                                {
                                  $ref: '#/components/schemas/YouTubeCategory',
                                },
                              type: array,
                            },
                        },
                    },
                },
            },
          tags: [YouTube],
          security: [],
          parameters:
            [
              {
                in: query,
                name: countryCode,
                required: false,
                schema: { default: US, type: string },
              },
            ],
        },
    }
  /youtube/videos:
    {
      get:
        {
          operationId: GetVideos,
          responses:
            {
              '200':
                {
                  description: '',
                  content:
                    {
                      application/json:
                        {
                          schema:
                            {
                              items:
                                { $ref: '#/components/schemas/YouTubeVideo' },
                              type: array,
                            },
                        },
                    },
                },
            },
          tags: [YouTube],
          security: [],
          parameters:
            [
              {
                in: query,
                name: countryCode,
                required: false,
                schema: { default: '', type: string },
              },
              {
                in: query,
                name: categoryId,
                required: false,
                schema: { default: '', type: string },
              },
              {
                in: query,
                name: trending,
                required: false,
                schema: { default: true, type: boolean },
              },
              {
                in: query,
                name: maxResults,
                required: false,
                schema: { default: 50, format: double, type: number },
              },
            ],
        },
    }
servers:
  - { url: 'http://localhost:8080/api', description: Local }
  - { url: 'http://dashboardworldapi.vercel.app/api', description: Vercel }